+++ freight/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = 'freight.apps.FreightConfig'
 
-__version__ = '0.11.3'
+__version__ = '0.11.4'
 __title__ = 'Alliance Freight'